Set-5/Prac3/requirements.txt
# Prac3 JWT Authentication - install with: pip install -r requirements.txt
Flask==3.0.0
PyJWT==2.8.0
